cleanup concurrency
